Simple CrudApp Project1Level1 CORE PHP
Did all the validation except the unique email thing.. The main thing took time was the pagination next button as the loop variable's last value was affecting the variable of the next button of the pagination. Hence, did the next and previous thing before the loop runs. Did the validation in front and backend.