Email Sending Functionalities
Used PHPMailer function for sending the email. It took some time as there was an error and somehow the PHPMailer was not being detected. Hence I uninstalled the composer and intalled and also added Vendor folder inside the same folder instead of getting it outside. Now next task is to verify the email with a link